backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django>=4.2.0
djangorestframework>=3.14.0
mysqlclient>=2.1.1
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
Pillow>=10.0.0
whitenoise>=6.5.0
google-generativeai>=0.3.0
django-redis>=5.3.0
django-rq>=2.8.1
bcrypt>=4.0.1.0
twilio>=8.10.0
sqlparse==0.4.4
tzdata==2024.1
gunicorn==21.2.0
whitenoise==6.6.0
